Setup recipe for configuring news and implement various templates
When enabling the news module, all the content definitions related to
news will be setup, a news content item that will hold all the news
posts and an admin menu option will be added.

In a seperate recipe that can be run manually, an example news post can
be added.