Update thumbnail defaults
So we don't recommend/enforce a small size for something which usually ends up being displayed very large. Also standardised behaviour a bit with how the Card widget works in terms of markup. I've also removed the bit that makes it say Draft as that seems to not be useful? But that could go back in if we want to keep it.